74 - [Blog] Style blog pages
