+++ styles/styles.css
@@ -253,6 +253,7 @@ em strong {
   font-size: var(--body-font-size-xs);
   display: block;
   color: var(--text-gray);
+  font-style: italic;
 }
 
 /* buttons */
@@ -653,7 +654,6 @@ main .section.wave {
   padding-bottom: 0;
   color: var(--text-white);
   text-align: center;
-  margin-bottom: 45px;
 }
 
 main .section.wave::after {
